Implemented pipe movement

+++ src/framebuffer.h
@@ -15,7 +15,7 @@ void buffer_initialise();
 void buffer_clear();
 bool buffer_get_pixel_state(uint8_t x, uint8_t y);
 void buffer_set_pixel_state(uint8_t x, uint8_t y, bool state);
-void buffer_draw_rectangle(uint8_t x1, uint8_t y1, uint8_t x2, uint8_t y2);
+void buffer_draw_rectangle(int x1, int y1, int x2, int y2);
 uint buffer_write_bitmap(uint8_t *bitmap, uint8_t height, int x, int y, bool erase);
 
 uint8_t _buffer_get_pixel_page_no(uint8_t y);
